Extract ReadReceipt into a separate file
